Bug fix for timezone. Reorganization of code
